refactor: move rate formatter to shared
